+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Next Release
 
 * Added support for the `~/.aws/credentials` INI file and credential profiles (via the `profile` option) as a safer
   alternative to using explicit credentials with the `key` and `secret` options.
+* Added support for query filters and improved conditional expressions to the Amazon DynamoDB client.
 * Added support for the `ChefConfiguration` parameter to a few operations on the AWS OpsWorks Client.
 * Added support for the `PlacementTenancy` parameter to the `CreateLaunchConfiguration` operation of the Auto Scaling
   client.